BUGGED

+++ src/main/java/model/BookStoreManagementRepository.java
@@ -3,5 +3,5 @@ package model;
 import org.springframework.data.repository.CrudRepository;
 
 public interface BookStoreManagementRepository extends CrudRepository<BookStoreManagement, Integer>{
-    BookStoreManagement findById(int id);
+    BookStoreManagement findById(long id);
 }
