+++ src/main/java/model/CustomerRepository.java
@@ -2,8 +2,8 @@ package model;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface CustomerRepository extends CrudRepository<Customer, Integer> {
-    Customer findById(int id);
+public interface CustomerRepository extends CrudRepository<Customer, Long> {
+    Customer findById(long id);
 
 }
